% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{build_schema}
\alias{build_schema}
\alias{build_schema.documentation}
\alias{build_schema.data.frame}
\title{Create JSON Schemas from R data frames and variable documentation objects.}
\usage{
build_schema(x, ...)

\method{build_schema}{documentation}(
  x,
  schema_key = "http://json-schema.org/draft-07/schema#",
  id_key = NULL,
  title_key = NULL,
  description_key = NULL,
  description_extras = "coding",
  as_schema = FALSE,
  ...
)

\method{build_schema}{data.frame}(
  x,
  schema_key = "http://json-schema.org/draft-07/schema#",
  id_key = NULL,
  title_key = NULL,
  description_key = NULL,
  description_extras = "coding",
  as_schema = FALSE,
  ...
)
}
\arguments{
\item{x}{a data frame or a documentation object.}

\item{...}{extra arguments passed to the methods; in case of \code{data.frame}
input those arguments are passed to \code{\link{create_doc}}.}

\item{schema_key}{value of the \verb{$schema} keyword in the JSON Schema header;
usually refers to the validation schema used later by a validator function.}

\item{id_key}{value of the \verb{$id} keyword in the JSON Schema header;
usually a project or table identifier.}

\item{title_key}{value of the \code{title} keyword in the JSON Schema header; usually
a project or table title}

\item{description_key}{value of the \code{description} keyword in the JSON Schema
header; usually a short description of a table or a project}

\item{description_extras}{names of the documentation columns to be included
in the variable description (\code{description} keyword value for variable
properties). Content of such column will be included in the description after
the \code{|} separator.}

\item{as_schema}{logical, should a \code{schema} class object be returned?
See Details.}
}
\value{
a string of class \code{\link{schema_string}} with the JSON Schema
structure (\code{as_schema = FALSE}) or a \code{\link{schema}} class object
(\code{as_schema = TRUE}).
}
\description{
\code{build_schema()} methods create JSON Schemas for R data frames, or variable
documentation objects generated by the user per hand or with
\code{\link{create_doc}}.
}
\details{
If \code{as_schema = TRUE}, the function returns a \code{\link{schema}} class
object generated from the JSON Schema string. Otherwise a plain string of
class \code{\link{schema_string}} is returned.
Those objects can be saved as JSON files on the disc with
\code{\link{write_schema}}.
}
\examples{
car_documentation <- create_doc(my_cars)
build_schema(car_documentation)

}
