% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{character_coding}
\alias{character_coding}
\title{Enumerate elements of a character vector.}
\usage{
character_coding(x)
}
\arguments{
\item{x}{a character vector.}
}
\value{
a string with the following numbers of elements and the elements of \code{x}
the \code{number: element} form separated by semicolons
}
\description{
Creates a string with enumerated unique elements of the character vector.
}
