% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_date}
\alias{check_date}
\title{Check if a vector contains dates.}
\usage{
check_date(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a logical value.
}
\description{
checks if a vector consists of dates.
}
