% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{create_coding}
\alias{create_coding}
\alias{create_coding.documentation}
\alias{create_coding.data.frame}
\title{Coding scheme of variables.}
\usage{
create_coding(x, ...)

\method{create_coding}{documentation}(x, safely = FALSE, ...)

\method{create_coding}{data.frame}(x, safely = FALSE, ...)
}
\arguments{
\item{x}{a \code{\link{documentation}} object or a data frame.}

\item{...}{extra arguments passed to \code{\link{create_doc}}.}

\item{safely}{logical: should any errors during parsing of the coding
strings be ignored? Possible parsing problems are returned as warnings.}
}
\value{
a data frame with the following columns:
\itemize{
\item \code{variable}: name of the variable
\item \code{description}: variable description extracted from the
\code{documentation} object
\item \code{value}: value assigned to the variable category
\item \code{label}: label of the variable category
}

If the \code{documentation} object or the data frame has no coding information,
the function return \code{NULL} with a warning.
}
\description{
The function generates a data frame with the variable coding scheme from a
\code{\link{documentation}} object or a data frame.
}
\details{
The coding information is extracted from the \code{coding} column of the
\code{\link{documentation}} object, which contains the a string of
\code{value: label} pairs separated by semicolons.
}
