% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{df2json}
\alias{df2json}
\title{Row-wise JSON data lists and JSON data strings from a data frame.}
\usage{
df2json(
  x,
  names_from = NULL,
  json_factor = c("integer", "string"),
  json_date = c("date-time", "date"),
  as_list = FALSE,
  ...
)
}
\arguments{
\item{x}{a data frame}

\item{names_from}{name of a variable in \code{x}, which specifies a unique
identifier used as names of the output list. If \code{names_from = '.rownames'},
row names of x will be used. If \code{names_from = NULL}, the output list will
have no numbers}

\item{json_factor}{specifies how factors of the data frame will be saved
in JSON data objects; by default as integers}

\item{json_date}{a string that specifies JSON date format, currently either
\code{date-time} (ISO 8601, default) or \code{date}}

\item{as_list}{logical, should JSON data lists be returned?
If \code{as_list = FALSE}, a list of JSON data strings is returned}

\item{...}{arguments passed to \code{\link[jsonlite]{toJSON}}}
}
\value{
a list of class \code{\link{json_data}} with JSON data strings or
JSON data lists.
}
\description{
Function \code{df2json} Converts a data frame into a list of JSON data stings or
JSON data lists in a row-wise manner, i.e. each element of the output list
represents a row of the data frame.
}
\details{
The function will run in parallel if a parallel backend is declared with
\code{\link[future]{plan}}. The backend will be closed afterwards.
}
\examples{
df2json(my_cars, names_from = 'ID')


}
