% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enumeration_string}
\alias{enumeration_string}
\title{Create enumeration for a numeric vector.}
\usage{
enumeration_string(x, enum_limit)
}
\arguments{
\item{x}{an atomic vector.}

\item{enum_limit}{an integer representing the enumeration limit.}
}
\value{
A string of unique values separated by commas if the length of
unique values is less than the enumeration limit, or \code{NULL} otherwise.
}
\description{
This function takes a vector and an enumeration limit.
If the vector has fewer unique values than the enumeration limit,
the function returns the a string of unique values.
Otherwise, it returns \code{NULL}. For factor vectors, the enumeration is
always returned.
}
