% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_utils.R
\name{format_markdown}
\alias{format_markdown}
\alias{format_html}
\title{Formats of strings in markdown and HTML.}
\usage{
format_markdown(
  x,
  format = c("plain", "italic", "bold", "bold.italic", "inline.code", "code")
)

format_html(
  x,
  format = c("plain", "italic", "bold", "bold.italic", "inline.code", "code"),
  div = TRUE
)
}
\arguments{
\item{x}{a character vector}

\item{format}{see \strong{Details} for available format options}

\item{div}{logical, should the elements of \code{x} be wrapped into \verb{<div>} tags?}
}
\value{
a character vector
}
\description{
Italic, bold, bold-italic, and code-style font face in Markdown- and
HTML-compatible format.
}
\details{
Options for \code{format} argument are:
\itemize{
\item \code{'plain'}: plain font face
\item \code{'italic'}: italic font face
\item \code{'bold'}: bold font face
\item \code{'bold.italic'}: bold-italic font face
\item \code{'code'}: the content will be displayed as code
}
}
