% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{get_stats}
\alias{get_stats}
\title{Descriptive/exploratory statistics for a data frame.}
\usage{
get_stats(
  x,
  variables = names(x),
  as_factor = TRUE,
  enum_limit = 10,
  signif_digits = 2,
  ...
)
}
\arguments{
\item{x}{a R data frame.}

\item{variables}{a character vector with names of the variables to
be analyzed. Defaults to all variables in the data frame.}

\item{as_factor}{should character variables be returned to factors prior
to computation of the statistics? Defaults to \code{TRUE}.}

\item{enum_limit}{the maximum of unique values of the variables, which turns
it into an enumerated variable.}

\item{signif_digits}{number of significant digits used for rounding of
the statistics.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with the variable names (column \code{variable}),
format (numeric, character, factor, date, column \code{format}) and descriptive
statistics (column \code{statistic}, see Details).
}
\description{
Function \code{get_stats()} computes descriptive statistics for columns of a data
frame.
}
\details{
For factors, enumerated character strings and enumerated numerical variables,
percentages and counts of observations within the categories are returned
along with the total numbers of complete observations.
For numerical variables and dates, medians with interquartile ranges and
ranges are computed.
}
