% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{json_data}
\alias{json_data}
\alias{is_json_data}
\title{JSON data list.}
\usage{
json_data(x, ...)

is_json_data(x)
}
\arguments{
\item{x}{a list with JSON data strings of JSON lists.}

\item{...}{extra arguments, currently none.}
}
\value{
a list of JSON data strings or JSON data lists.
}
\description{
A container list for JSON data strings or lists.
}
