% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_utils.R
\name{mapColSpecs}
\alias{mapColSpecs}
\title{Convert JSON Schema variable properties to column-defining SQL expressions.}
\usage{
mapColSpecs(x, name, required = NULL, char_size = 255)
}
\arguments{
\item{x}{a list of variable (column) properties corresponding to the
JSON Schema key: value pairs.}

\item{name}{variable name.}

\item{required}{a character vector with names of required variables}

\item{char_size}{size of th character strings}
}
\value{
a string with SQL expressions that define table columns.
}
\description{
Takes variable properties specified within the \code{properties} key of a
JSON Schema and turns them into SQL expressions that define a table column.
}
