% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_utils.R
\name{mapDataType}
\alias{mapDataType}
\title{Map JSON Schema variable types to SQL data types.}
\usage{
mapDataType(type, char_size = 255)
}
\arguments{
\item{type}{a character string with the JSON Data type.}

\item{char_size}{size of th character strings}
}
\value{
a string with the SQL data type.
}
\description{
Maps the JSON Schema variable types to SQL data types.
}
