% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_cars}
\alias{my_cars}
\title{\code{my_cars}: a derivative of the Cars93 Data Set}
\format{
A data frame with 93 rows and 31 variables:
\itemize{
\item \code{ID}: unique identifier of the car
\item \code{Manufacturer}: Manufacturer of the car
\item \code{Model}: Model of the car
\item \code{Type}: Type of car (e.g., Small, Midsize, Large, etc.)
\item \code{Min.Price}: Minimum price of the car in thousands of dollars
\item \code{Price}: Midrange price of the car in thousands of dollars
\item \code{Max.Price}: Maximum price of the car in thousands of dollars
\item \code{MPG.city}: Miles per gallon in city driving
\item \code{MPG.highway}: Miles per gallon on the highway
\item \code{AirBags}: Type of airbags (None, Driver only, Driver & Passenger)
\item \code{DriveTrain}: Type of drivetrain (Rear, Front, 4WD)
\item \code{Cylinders}: Number of cylinders
\item \code{EngineSize}: Engine size in liters
\item \code{Horsepower}: Horsepower of the car
\item \code{RPM}: Revolutions per minute at which maximum horsepower is achieved
\item \code{Rev.per.mile}: Revolutions per mile
\item \code{Man.trans.avail}: Availability of manual transmission (Yes or No)
\item \code{Fuel.tank.capacity}: Fuel tank capacity in gallons
\item \code{Passengers}: Number of passengers
\item \code{Length}: Length of the car in inches
\item \code{Wheelbase}: Wheelbase in inches
\item \code{Width}: Width of the car in inches
\item \code{Turn.circle}: Turning circle in feet
\item \code{Rear.seat.room}: Rear seat room in inches
\item \code{Luggage.room}: Luggage room in cubic feet
\item \code{Weight}: Weight of the car in pounds
\item \code{Origin}: Country of origin (USA or non-USA)
\item \code{Make}: Manufacturer and model combined provided as string
\item \code{Origin_free_text}: a sting version of the \code{Origin} variable
\item \code{Cylinder_number}: number of cylinders specified as an integer
\item \code{Entry}: a fictional date of entry in the data set
}
}
\source{
\url{https://vincentarelbundock.github.io/Rdatasets/doc/MASS/Cars93.html}
}
\usage{
my_cars
}
\description{
This data set contains information about 93 cars from the 1993 model year.
The data includes a variety of variables related to the characteristics,
performance, and pricing of the cars.
}
\examples{
data(my_cars)
head(my_cars)
summary(my_cars)
}
\keyword{datasets}
