% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_unique}
\alias{n_unique}
\title{Counts of unique elements in a vector.}
\usage{
n_unique(x)
}
\arguments{
\item{x}{an atomic vector.}
}
\value{
an integer with the number of unique elements.
}
\description{
Counts of unique elements in an atomic vector.
}
