% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_stats}
\alias{num_stats}
\title{Statistics for numeric features and dates.}
\usage{
num_stats(x, as_vector = FALSE, signif_digits = 2)
}
\arguments{
\item{x}{a numeric or date (Date, POSIXct, POSIXt, POSIXlt) vector}

\item{as_vector}{logical, should a named vector of counts, percentages, or
percentages and counts be returned?}

\item{signif_digits}{number of significant digits to round the function's
output. Use only if \code{as_vector = FALSE}}
}
\value{
if \code{as_vector = TRUE} a numeric vector with the statistics is
returned, otherwise a pre-formatted sting with the statistic names
and values.
}
\description{
The function computes medians, interquartile ranges, and ranges for numeric
vectors and dates.
}
