% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_coding}
\alias{parse_coding}
\title{Coding scheme from a coding string.}
\usage{
parse_coding(x, as_data_frame = FALSE, safely = FALSE)
}
\arguments{
\item{x}{a character string as described in \strong{Description}.}

\item{as_data_frame}{logical: a data frame output?}

\item{safely}{logical. If \code{safely = TRUE}, the function returns \code{NULL}
with a warning if any parsing errors are encountered. Otherwise, an
error is raised.}
}
\value{
a list of \code{c(value, label)} tuples or a data frame with
\code{value} and \code{label} columns.
}
\description{
The function takes a single character string of \code{value: label} pairs
separated by semicolons and returns a list of \code{c(value, label)} tuples or
a data frame with \code{value} and \code{label} columns.
}
\details{
\code{value} is considered as a usually numeric value returned at the data base
side, while \code{label} is meant as the label presented at the user's interface
(for example in forms and views).
}
