% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{render_doc}
\alias{render_doc}
\alias{render_doc.documentation}
\alias{render_doc.data.frame}
\title{Render table documentation as markdown or HTML code chunks.}
\usage{
render_doc(x, ...)

\method{render_doc}{documentation}(
  x,
  type = c("markdown", "html"),
  col_include = NULL,
  col_labels = NULL,
  heading_levels = c(1, 2),
  ...
)

\method{render_doc}{data.frame}(
  x,
  type = c("markdown", "html"),
  col_include = NULL,
  col_labels = NULL,
  heading_levels = c(1, 2),
  ...
)
}
\arguments{
\item{x}{a data frame or a \code{documentation} object.}

\item{...}{extra parameters passed to \code{\link{create_doc}} and methods.}

\item{type}{type of the generated code: markdown (default) or HTML.}

\item{col_include}{columns of the \code{\link{documentation}} object to be
rendered. By default, i.e. \code{NULL}, all columns in the documentation object
are rendered.}

\item{col_labels}{labels for the columns \code{\link{documentation}} object
which will be displayed as headers in the code. Defaults to \code{NULL},
which causes the documentation columns \code{variable}, \code{type_r}, \code{enumeration},
\code{coding}, \code{description}, \code{json_expr} and \code{required} to be displayed as,
respectively, 'Variable name', 'Format', 'Enumeration', 'Coding scheme',
'Description', 'JSON Schema rules' and 'Required'.
With this default option, any extra columns in the documentation object will
get captions corresponding to their plain names.}

\item{heading_levels}{an integer tuple that defines markdown or HTML heading
levels. Defaults to \code{c(1, 2)}, which means that the caption of the chunk with
the variable name will be a \code{h1} header, and sub-captions will be displayed
as \code{h2} headers.}
}
\value{
A data frame of class \code{\link{renDoc}} with the following columns:
\itemize{
\item \code{variable} with name of the variable to be documented
\item \code{code_type} which indicates the type of generated code (markdown or HTML)
\item \code{code} with the markdown or HTML codem with variable documentation
}

See also \code{\link{toDocument.renDoc}} method that can be used to generate
ready-to-use Markdown and HTML documents with documentation of the data
frame's variables.
}
\description{
Function \code{render_doc()} generates chunks of Markdown or HTML code based on
documentation of variables in R data frames.
Such chunks may be easily used to create user-friendly documantation and
help pages for the data frame's metadata.
}
