% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{renDoc}
\alias{renDoc}
\alias{is_renDoc}
\title{\code{renDoc} objects holding markdown or HTML chunks with variable documentation.}
\usage{
renDoc(x, ...)

is_renDoc(x)
}
\arguments{
\item{x}{a data frame with the with columns described in \strong{Details}.}

\item{...}{extra arguments, currently none.}
}
\value{
A data fame with of class \code{renDoc} with columns described in \strong{Details}.
}
\description{
\code{renDoc} class objects are generated by \code{\link{render_doc}} function
and hold markdown or HTML code with documentation of variables of a R data
frame.
}
\details{
\code{renDoc} objects are data frames with the following columns:
\itemize{
\item \code{variable} with name of the variable to be documented
\item \code{code_type} which indicates the type of generated code (markdown or HTML)
\item \code{code} with the markdown or HTML code with variable documentation
}

\code{renDoc} data frames can be transformed like usual data frames.
See also \code{\link{toDocument.renDoc}} method that can be used to generate
ready-to-use Markdown and HTML documents with documentation of the data
frame's variables.
}
