% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{row2json}
\alias{row2json}
\title{Generate a JSON string from a row (record) of a data frame.}
\usage{
row2json(x, idx, as_list = FALSE, ...)
}
\arguments{
\item{x}{a data frame}

\item{idx}{a row index.}

\item{as_list}{logical, should JSON data lists be returned?
If \code{as_list = FALSE}, a list of JSON data strings is returned.}

\item{...}{arguments passed to \code{\link[jsonlite]{toJSON}}.}
}
\value{
a JSON string generated by \code{\link[jsonlite]{toJSON}}
}
\description{
Converts a row of a data frame with the given index into a JSON string.
}
