% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_utils.R
\name{row2markdown}
\alias{row2markdown}
\alias{row2html}
\title{Markdown and HTML code for a row of a data frame.}
\usage{
row2markdown(x, title = NULL, format = NULL, heading_levels = c(1, 2))

row2html(x, title = NULL, format = NULL, heading_levels = c(1, 2))
}
\arguments{
\item{x}{a data frame. Only the first row will be processed.}

\item{title}{text to be displayed as the chunk title. Defaults to the content
of the first column of \code{x}}

\item{format}{\code{NULL} or an optional character vector with the length that
equals the number of columns in \code{x} and specifies the formatting of the
content. See \strong{Details} for available formatting options.}

\item{heading_levels}{an integer tuple that defines markdown or HTML heading
levels. Defaults to \code{c(1, 2)}, which means that the caption of the chunk with
the variable name will be a \code{h1} header, and sub-captions will be displayed
as \code{h2} headers.}
}
\description{
Generates Markdown (\code{row2markdown()}) or HTML code (\code{row2html}) for the
column names as headers and the column content as paragraphs.
}
\details{
Intended for internal use.

Options for \code{format} argument are:
\itemize{
\item \code{'italic'}: italic font face
\item \code{'bold'}: bold font face
\item \code{'bold.italic'}: bold-italic font face
\item \code{'code'}: the content will be displayed as code
}
}
