% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{schema}
\alias{schema}
\alias{is_schema}
\title{JSON Schema lists.}
\usage{
schema(x, ...)

is_schema(x)
}
\arguments{
\item{x}{a list with elements specified in Details.}

\item{...}{extra arguments, currently none.}
}
\value{
a list of class \code{schema} with the elements specified in Details.
}
\description{
Creates an instance of \code{schema} class, which is a minimal representation of
a \href{https://json-schema.org/}{JSON Schema}.
}
\details{
The function returns a list with the following elements:
\itemize{
\item \verb{$schema}: reference to the \href{https://json-schema.org/}{JSON Schema} version
used by validator functions; this is a part of the JSON Schema header
\item \verb{$id}: an object id, usually an identifier of a table or a project; this
is a part of the JSON Schema header
\item \code{title}: a title of the schema, usually a title of a table or a project;
this is a part of the JSON Schema header
\item \code{description}: a description of the schema, usually a description of a
table or a project; his is a part of the JSON Schema header
\item \code{type}: the value of the \code{type} keyword in the JSON Schema header; usually
set to \code{"type": "object"}
\item \code{properties}: a list of variable properties such as type, enumeration,
description, minimum, maximum, and so on
\item \code{required}: an optional element which specifies the required variables
}

See \code{\link{write_schema}} for a function for saving the JSON Schema
list as a JSON file on the disc.
}
