% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_constructors.R
\name{schema2sql}
\alias{schema2sql}
\alias{schema2sql.schema_string}
\alias{schema2sql.schema}
\title{Create a SQL table given a JSON Schema.}
\usage{
schema2sql(x, ...)

\method{schema2sql}{schema_string}(x, beautiful = TRUE, ...)

\method{schema2sql}{schema}(x, beautiful = TRUE, ...)
}
\arguments{
\item{x}{a \code{\link{schema}} or \code{\link{schema_string}} object}

\item{...}{extra arguments passed to methods}

\item{beautiful}{logical, should line breaks be introduced?}
}
\value{
a character string containing a \verb{CREATE TABLE} SQL expression.
}
\description{
\code{schema2sql()} methods generate SQL expressions that create a table given
a \href{https://json-schema.org/}{JSON Schema}.
}
\details{
The methods are specified for \code{\link{schema}} and
\code{\link{schema_string}} objects. Note: tables can be created only from
JSON Schemes with a non-empty keyword \verb{$id}; the \verb{$id} keyword serves as the
table name.
}
