% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{schema_string}
\alias{schema_string}
\alias{is_schema_string}
\title{JSON Schema strings.}
\usage{
schema_string(x, ...)

is_schema_string(x)
}
\arguments{
\item{x}{a string.}

\item{...}{extra arguments, currently none.}
}
\value{
a string of class \code{json_schema}.
}
\description{
Creates an instance of \code{schema_string}: a string that contains a complete
minimal \href{https://json-schema.org/}{JSON Schema}.
}
\details{
See \code{\link{write_schema}} for a function for saving the JSON Schema
string as a JSON file on the disc.
}
