% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_search.R
\name{show_tags}
\alias{show_tags}
\alias{add_tags}
\alias{delete_tags}
\alias{filter_tags}
\alias{expand_tags}
\title{Display, add, remove, and search variable tags in data frame documentation.}
\usage{
show_tags(x, ...)

add_tags(x, tags, ...)

delete_tags(x, tags, ...)

filter_tags(x, tags, mode = c("any", "all"))

expand_tags(x, tags, one_column = FALSE)
}
\arguments{
\item{x}{a \code{documentation} object.}

\item{...}{one or more logical statements used for selection of variables in
the \code{documentation} object. See \code{\link[dplyr]{filter}} for details.}

\item{tags}{a character vector with the tags.}

\item{mode}{full-text search mode for filtering variable tags.
See the \strong{Details}.}

\item{one_column}{should names of exclusive tags be placed in one column?
See the \strong{Details}.}
}
\value{
a \code{\link{documentation}} object with column \code{tags} storing
character vectors of tags.
}
\description{
Non-hierarchical and hierarchical tags are a convenient way to classify
metadata e.g. according to their semantics or quality.

\code{show_tags()}, \code{add_tags()} and \code{delete_tags()} allow showing, adding and
removing user-specified tags of variables selected by one or more logical
conditions.
For details on storage and modification of tags, please refer to \strong{Details}.

\code{filter_tags()} filters the documentation object by presence of tags.
\code{expand_tags()} expands tags of the documentation object into one or more
separate columns as described in \strong{Details}.
}
\details{
Technically, tags are stored as a list of character vectors in the \code{tags}
column of a \code{documentation} object.
Variables whose tags will be modified are selected by logical statements
provided via \code{...} argument of the \code{show_tags()} ,\code{add_tags()}, and
\code{delete_tags()} functions.
In general, this semantics copies the semantics of
\code{\link[dplyr]{filter}}.
If no statements are provided to \code{...}, all tags in the \code{documentation}
object are displayed or altered.

\code{filter_tags()} executes full-text search for tags.
If \code{mode = 'any'}, the rows are returned, where any of the searching tags
specified by \code{tags} matches the vaiable tags.
If \code{mode = 'all'}, a full march between the \code{tags} argument and the variable
tags is expected.

\code{expand_tags()} converts the tags in a documentation object specified by
\code{tags} argument into one or more columns named after these tags of interest.
if \code{one_column = FALSE} each of these columns contains \code{TRUE/FALSE} logical
values indicating for the tag presence.
if \code{one_column = TRUE} and the tags of interest are exclusive, one character
column \code{tag_class} with the tag names is added.
}
