% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{string2schema}
\alias{string2schema}
\alias{schema2string}
\title{Conversion between \code{schema} objects with JSON Schemes and JSON strings.}
\usage{
string2schema(x, ...)

schema2string(x, ...)
}
\arguments{
\item{x}{a \code{\link{schema}} or \code{\link{schema_string}} object.}

\item{...}{additional arguments, currently none.}
}
\value{
\code{string2schema()} returns a JSON Schema list (class \code{\link{schema}}),
\code{schema2string()} returns a JSON Schema string (class
\code{\link{schema_string}}).
}
\description{
The functions \code{string2schema()} and \code{schema2string()} enable conversion
between JSON Schema strings (class \code{\link{schema_string}}) and
JSON Schema lists (class \code{\link{schema}}).
}
