% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tab}
\alias{tab}
\title{Tabulate a vector: calculate frequency of elements.}
\usage{
tab(
  x,
  type = c("counts", "percents", "both"),
  as_vector = TRUE,
  signif_digits = 2
)
}
\arguments{
\item{x}{an atomic vector.}

\item{type}{type of the output: counts, percentages, or percentages and
counts. Defaults to counts and is ignored if \code{as_vector = FALSE}.}

\item{as_vector}{logical, should a named vector of counts, percentages, or
percentages and counts be returned?}

\item{signif_digits}{number of significant digits to round the percentages.
Used only if \code{as_vector = FALSE} or \code{type = 'both'}.}
}
\value{
If \code{as_vector = TRUE}, the function returns a vector with counts, percentages,
or percentages and counts (general form: percentage \% (n = )).
If \code{as_vector = FALSE}, the function returns a single string with percentages,
counts, and numbers of complete cases.
}
\description{
Function \code{tab()} computes frequencies and, optionaly, percentages of
elements of a vector.
}
