% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{toDocument}
\alias{toDocument}
\alias{toDocument.renDoc}
\alias{toDocument.documentation}
\title{Markdown and HTML documents with variable documentation.}
\usage{
toDocument(x, ...)

\method{toDocument}{renDoc}(
  x,
  title = "Variables",
  subtitle = "Variable documentation",
  sep = "<hr>",
  file = NULL,
  ...
)

\method{toDocument}{documentation}(
  x,
  title = "Variables",
  subtitle = "Variable documentation",
  sep = "<hr>",
  file = NULL,
  ...
)
}
\arguments{
\item{x}{a \code{renDoc} or \code{documentation} object.}

\item{...}{additional arguments passed to \code{\link{render_doc}}
and methods.}

\item{title}{a character string to be used as the document's title.}

\item{subtitle}{a character string to be used as the document's subtitle.}

\item{sep}{separator between the variable documentation chunks.}

\item{file}{a optional argument that specifies path of the markdown ('.md')
or HTML file ('.html'), or a connection to write to.}
}
\value{
\code{toDocument()} method returns a single character string that represents the
entire markdown or HTML document with the variable documentation.
If a file path is provided as \code{file} argument, the document is additionally
saved on a disc.
}
\description{
\code{toDocument()} applied to \code{\link{documentation}} or \code{\link{renDoc}}
objects, the later created with \code{\link{render_doc}} function, generates
ready-to-use markdown or HTML documents with documentation of a data frame's
variables. Such markdown or HTML documents may be used e.g. as help pages or
fo construction of metadata catalogs.
}
