% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{write_schema}
\alias{write_schema}
\alias{write_schema.schema_string}
\alias{write_schema.schema}
\title{Writing JSON Schemas on the disc.}
\usage{
write_schema(x, path, ...)

\method{write_schema}{schema_string}(x, path, ...)

\method{write_schema}{schema}(x, path, ...)
}
\arguments{
\item{x}{a \code{\link{schema_string}} or \code{\link{schema}} object.}

\item{path}{a path to the file.}

\item{...}{extra arguments passed to the methods.}
}
\value{
returns invisibly the path to the JSON file.
}
\description{
The \code{write_schema} methods save JSON Schemas as JSON files on the disc.
}
