% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{build_schema}
\alias{build_schema}
\alias{build_schema.documentation}
\alias{build_schema.data.frame}
\title{Create JSON Schemas from R data frames and variable documentation objects.}
\usage{
build_schema(x, ...)

\method{build_schema}{documentation}(
  x,
  schema_key = "http://json-schema.org/draft-07/schema#",
  id_key = NULL,
  title_key = NULL,
  description_key = NULL,
  description_extras = "coding",
  extras_keywords = FALSE,
  tail_statements = NULL,
  as_schema = FALSE,
  ...
)

\method{build_schema}{data.frame}(
  x,
  schema_key = "http://json-schema.org/draft-07/schema#",
  id_key = NULL,
  title_key = NULL,
  description_key = NULL,
  description_extras = "coding",
  extras_keywords = FALSE,
  tail_statements = NULL,
  as_schema = FALSE,
  ...
)
}
\arguments{
\item{x}{a data frame or a documentation object.}

\item{...}{extra arguments passed to the methods; in case of \code{data.frame}
input those arguments are passed to \code{\link{create_doc}}.}

\item{schema_key}{value of the \verb{$schema} keyword in the JSON Schema header;
usually refers to the validation schema used later by a validator function.}

\item{id_key}{value of the \verb{$id} keyword in the JSON Schema header;
usually a project or table identifier.}

\item{title_key}{value of the \code{title} keyword in the JSON Schema header; usually
a project or table title}

\item{description_key}{value of the \code{description} keyword in the JSON Schema
header; usually a short description of a table or a project}

\item{description_extras}{names of the documentation columns to be included
in the variable description (\code{description} keyword value for variable
properties). At the moment variable tags (\code{tags} column of documentation)
will not included in the schema, event if listed here.}

\item{extras_keywords}{logical, should the documentation columns specified
by \code{description_extras} be included as separate keywords tethered to the
variable's properties? Defaults to \code{FALSE}, which means that content of such
column will be pasted with the description after the \code{|} separator. Please
note that extra keywords in variable are not necessarily compatible with
standard JSON Schema syntax without declaration of meta-data dictionaries.
For discussion, see:
https://json-schema.org/understanding-json-schema/reference/schema#guidelines}

\item{tail_statements}{an optional character string that specifies additional
validation keywords and their values such as
\href{https://json-schema.org/understanding-json-schema/reference/conditionals#dependentRequired}{dependent requirements}
or \href{https://json-schema.org/understanding-json-schema/reference/conditionals#dependentSchemas}{dependent schemas}}

\item{as_schema}{logical, should a \code{schema} class object be returned?
See Details.}
}
\value{
a string of class \code{\link{schema_string}} with the JSON Schema
structure (\code{as_schema = FALSE}) or a \code{\link{schema}} class object
(\code{as_schema = TRUE}).
}
\description{
\code{build_schema()} methods create JSON Schemas for R data frames, or variable
documentation objects generated by the user per hand or with
\code{\link{create_doc}}.
}
\details{
If \code{as_schema = TRUE}, the function returns a \code{\link{schema}} class
object generated from the JSON Schema string. Otherwise a plain string of
class \code{\link{schema_string}} is returned.
Those objects can be saved as JSON files on the disc with
\code{\link{write_schema}}.

Please note, that the output of \code{build_schema()} function provides only the
core validation functionality and does not allow for generation dependent
requirements or dependent schemas (sub-schemas) - because this information
cannot be extracted from a plain data frame.
Such extra validation statements may be specified via \code{tail_statement}
argument.

By specifying \code{description_extras}, the user may include additional columns
from the documentation in the JSON Schema.
If \code{extras_keywords = FALSE}, this extra portion information will be pasted
with the \code{description} keyword with the \code{|} separator.
If \code{extras_keywords = TRUE}, the additional columns, if not empty, will be
included in the schema as additional keywords under \code{property} key of the
variables.
Please note that such custom keywords may not be automatically recognized
by standard JSON Schema validators.
See: https://json-schema.org/understanding-json-schema/reference/schema#guidelines
}
\examples{
car_documentation <- create_doc(my_cars)
build_schema(car_documentation)

}
