% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_search.R
\name{show_tags}
\alias{show_tags}
\alias{add_tags}
\alias{delete_tags}
\alias{filter_tags}
\title{Display, add, remove, and search variable tags in data frame documentation.}
\usage{
show_tags(x, ...)

add_tags(x, tags, ...)

delete_tags(x, tags, ...)

filter_tags(x, tags, mode = c("any", "all"))
}
\arguments{
\item{x}{a \code{documentation} object.}

\item{...}{one or more logical statements used for selection of variables in
the \code{documentation} object. See \code{\link[dplyr]{filter}} for details.}

\item{tags}{a character vector with the tags.}

\item{mode}{full-text search mode for filtering variable tags.
See \strong{Details}.}
}
\value{
a \code{\link{documentation}} object with column \code{tags} storing
character vectors of tags.
}
\description{
Non-hierarchical and hierarchical tags are a convenient way to classify
metadata e.g. according to their semantics or quality.

\code{show_tags()}, \code{add_tags()} and \code{delete_tags()} allow showing, adding and
removing user-specified tags of variables selected by one or more logical
conditions.
For details on storage and modification of tags, please refer to \strong{Details}.

\code{filter_tags()} filters the documentation object by presence of tags.
}
\details{
Technically, tags are stored as a list of character vectors in the \code{tags}
column of a \code{documentation} object.
Variables whose tags will be modified are selected by logical statements
provided via \code{...} argument of the \code{show_tags()} ,\code{add_tags()}, and
\code{delete_tags()} functions.
In general, this semantics copies the semantics of
\code{\link[dplyr]{filter}}.
If no statements are provided to \code{...}, all tags in the \code{documentation}
object are displayed or altered.

\code{filter_tags()} executes full-text search for tags.
If \code{mode = 'any'}, rows are returned, where any of the searching tags
specified by \code{tags} matches the vaiable tags.
If \code{mode = 'all'}, a full march between the \code{tags} argument and the variable
tags is expected.
}
